authorization implemented

+++ src/services/notes/controllers/note-controller.js
@@ -3,14 +3,17 @@ import { nanoid } from "nanoid";
 import NoteRepositories from "../repositories/note-repositories.js";
 import { InvariantError, NotFoundError } from "../../../exceptions/index.js";
 import response from "../../../utils/response.js";
+import AuthorizationError from "../../../exceptions/authorization-error.js";
 
 //  menghilangkan try catch, jika error karena kesalahan logika, akan masuk ke InvariantError.
 export const createNote = async (req, res, next) => {
   const { title, body, tags } = req.validated;
+  const { id: owner } = req.user;
   const note = await NoteRepositories.createNote({
     title,
     body,
     tags,
+    owner,
   });
   if (!note) {
     return next(new InvariantError("Catatan gagal ditambahkan"));
@@ -19,13 +22,21 @@ export const createNote = async (req, res, next) => {
 };
 
 export const getNotes = async (req, res) => {
-  const notes = await NoteRepositories.getNotes();
+  const { id: owner } = req.user;
+  const notes = await NoteRepositories.getNotes(owner);
   return response(res, 200, "Catatan sukses ditampilkan", notes);
 };
 
 // Jika catatan tidak ditemukan, akan masuk ke NotFoundError.
 export const getNoteById = async (req, res, next) => {
   const { id } = req.params;
+  const { id: owner } = req.user;
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(
+      new AuthorizationError("Anda tidak berhak mengakses resource ini"),
+    );
+  }
   const note = await NoteRepositories.getNoteById(id);
   if (!note) {
     return next(new NotFoundError("Catatan tidak ditemukan"));
@@ -36,6 +47,15 @@ export const getNoteById = async (req, res, next) => {
 export const editNote = async (req, res, next) => {
   const { id } = req.params;
   const { title, body, tags } = req.validated;
+  const { id: owner } = req.user;
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+
+  if (!isOwner) {
+    return next(
+      new AuthorizationError("Anda tidak berhak mengakses resource ini"),
+    );
+  }
+
   const note = await NoteRepositories.editNote({
     id,
     title,
@@ -50,6 +70,13 @@ export const editNote = async (req, res, next) => {
 
 export const deleteNote = async (req, res, next) => {
   const { id } = req.params;
+  const { id: owner } = req.user;
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(
+      new AuthorizationError("Anda tidak berhak mengakses resource ini"),
+    );
+  }
   const deletedNote = await NoteRepositories.deleteNote(id);
   if (!deletedNote) {
     return next(new NotFoundError("Catatan tidak ditemukan"));
